sistema-codigo-barras/backend/requirements.txt
# Dependências base do Flask
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3

# Dependências específicas do projeto
requests==2.31.0          # Para requisições HTTP (API de CEP)
python-barcode==0.15.1    # Para geração de códigos de barras
Pillow==10.0.1            # Para processamento de imagens
python-dotenv==1.0.0      # Para carregar variáveis de ambiente
marshmallow==3.20.1       # Para serialização/validação de dados
python-dateutil==2.8.2    # Para manipulação de datas
uuid==1.30                # Para geração de IDs únicos

# Dependências para desenvolvimento e testes
pytest==7.4.2            # Framework de testes
pytest-flask==1.2.0      # Extensão do pytest para Flask
coverage==7.3.2          # Cobertura de testes

